+++ grace-web-mvc/src/main/groovy/org/grails/web/controller/DefaultControllerExceptionHandlerMetaData.groovy
@@ -1,5 +1,5 @@
 /*
- * Copyright 2013-2022 the original author or authors.
+ * Copyright 2013-2026 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.grails.plugins.web.controllers
+package org.grails.web.controller
 
 import groovy.transform.Immutable
